requirements-full.txt
# Optional dependencies for local real_full DistilBERT + AG News, MobileNetV2 + CIFAR-10, and CLIP + Flickr30K / ViLT + Flickr30K pipelines
# Not installed in CI by default.
torch>=2.0
transformers>=4.40
torchvision>=0.15
pillow>=10.0
tqdm>=4.66
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
captum>=0.7

onnx>=1.16
onnxruntime>=1.18
